WiP: change installation flow of Firefox according to the official? one
According to the official page (and some related pages), the installation step for Firefox seems to differ from the lines described in this Dockerfile. I'm not sure, but there may have been a change in the official procedure.
https://support.mozilla.org/kb/install-firefox-linux
https://www.omgubuntu.co.uk/2022/04/how-to-install-firefox-deb-apt-ubuntu-22-04

However, the image built by this fix could not start Firefox app on the VNC though,,,